--- js/data/levels-data.js
@@ -1,84 +1,0 @@
-export const levels = {
-  'level-0': {
-    type: `artist`,
-    src: `https://www.youtube.com/audiolibrary_download?vid=91624fdc22fc54ed`,
-    answers: [
-      {
-        correct: true,
-        src: `http://placehold.it/134x134`,
-        artist: `Пелагея`
-      },
-      {
-        src: `http://placehold.it/134x134`,
-        artist: `The Prodigy`
-      },
-      {
-        src: `http://placehold.it/134x134`,
-        artist: `Lorde`
-      }
-    ]
-  },
-  'level-1': {
-    type: `genre`,
-    title: `Выберите инди-рок треки`,
-    questions: [
-      {
-        src: `https://www.youtube.com/audiolibrary_download?vid=91624fdc22fc54ed`,
-        correct: true
-      },
-      {
-        src: `https://www.youtube.com/audiolibrary_download?vid=dc3b4dc549becd6b`,
-        correct: false
-      },
-      {
-        src: `https://www.youtube.com/audiolibrary_download?vid=a127d9b7de8a17cf`,
-        correct: true
-      },
-      {
-        src: `https://www.youtube.com/audiolibrary_download?vid=dfb828f40096184c`,
-        correct: false
-      }
-    ]
-  },
-  'level-2': {
-    type: `artist`,
-    src: `https://www.youtube.com/audiolibrary_download?vid=91624fdc22fc54ed`,
-    answers: [
-      {
-        correct: true,
-        src: `http://placehold.it/134x134`,
-        artist: `Пелагея`
-      },
-      {
-        src: `http://placehold.it/134x134`,
-        artist: `The Prodigy`
-      },
-      {
-        src: `http://placehold.it/134x134`,
-        artist: `Lorde`
-      }
-    ]
-  },
-  'level-3': {
-    type: `genre`,
-    title: `Выберите инди-рок треки`,
-    questions: [
-      {
-        src: `https://www.youtube.com/audiolibrary_download?vid=91624fdc22fc54ed`,
-        correct: true
-      },
-      {
-        src: `https://www.youtube.com/audiolibrary_download?vid=dc3b4dc549becd6b`,
-        correct: false
-      },
-      {
-        src: `https://www.youtube.com/audiolibrary_download?vid=a127d9b7de8a17cf`,
-        correct: true
-      },
-      {
-        src: `https://www.youtube.com/audiolibrary_download?vid=dfb828f40096184c`,
-        correct: false
-      }
-    ]
-  }
-};
