+++ js/screens/artist.js
@@ -1,26 +1,6 @@
 import {createElementFromHTML} from '../utils';
-import {initialState} from '../data/game-data';
 import {headerHTML} from './header';
 import {playerHTML, onPlayerControlClick} from '../audio';
-
-const level = {
-  src: `https://www.youtube.com/audiolibrary_download?vid=91624fdc22fc54ed`,
-  answers: [
-    {
-      correct: true,
-      src: `http://placehold.it/134x134`,
-      artist: `Пелагея`
-    },
-    {
-      src: `http://placehold.it/134x134`,
-      artist: `The Prodigy`
-    },
-    {
-      src: `http://placehold.it/134x134`,
-      artist: `Lorde`
-    }
-  ]
-};
 
 const answerHTML = (item, i) => {
   return `
@@ -35,35 +15,38 @@ const answerHTML = (item, i) => {
   `;
 };
 
-const artistLevel = (state) => {
-  return createElementFromHTML(`
+const artistLevel = () => {
+  const gameState = window.gameState;
+  const currentLevel = window.currentLevel;
+
+  const node = createElementFromHTML(`
     <section class="main main--level main--level-artist">
-      ${headerHTML(initialState)}
+      ${headerHTML(gameState)}
       <div class="main-wrap">
         <h2 class="title main-title">Кто исполняет эту песню?</h2>
         <div class="player-wrapper">
-          ${playerHTML(`${state.src}`, false)}
+          ${playerHTML(`${currentLevel.questionSrc}`, false)}
         </div>
         <form class="main-list">
-          ${state.answers.map((answer, index) => answerHTML(answer, index)).join(``)}
+          ${currentLevel.answers.map((answer, index) => answerHTML(answer, index)).join(``)}
         </form>
       </div>
     </section>`);
+
+  node.querySelector(`.main-list`).addEventListener(`change`, (evt) => {
+    const answerIndex = +evt.target.value.split(`-`)[1];
+    if (currentLevel.answers[answerIndex].correct === true) {
+      window.correctAnswer();
+    } else {
+      window.wrongAnswer();
+    }
+  });
+
+  node.querySelector(`.player-control`).addEventListener(`click`, (evt) => {
+    onPlayerControlClick(evt);
+  });
+
+  return node;
 };
 
-const artistScreen = artistLevel(level);
-
-artistScreen.querySelector(`.main-list`).addEventListener(`change`, (evt) => {
-  const answerIndex = +evt.target.value.split(`-`)[1];
-  if (level.answers[answerIndex].correct === true) {
-    window.correctAnswer();
-  } else {
-    window.wrongAnswer();
-  }
-});
-
-artistScreen.querySelector(`.player-control`).addEventListener(`click`, (evt) => {
-  onPlayerControlClick(evt);
-});
-
-export default artistScreen;
+export default artistLevel;
